:recycle: refactored paypal to component :bug: fixed form submission
